third_party/blink/web_tests/inspector-protocol/dom/dom-svg-attribute-case-expected.txt

Test that DOM attribute case is preserved when modifying SVG elements.
Original attributes:
id=main
xmlns=http://www.w3.org/2000/svg
width=600
height=500
viewBox=0 0 100 120
Modified attribute:
viewBox=0 0 120 120
